requirements.txt
#sudo apt-get install ffmpeg
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
# ecdsa
# paramiko
# pycrypto
flask
azure-cognitiveservices-speech
azure-ai-textanalytics
openai-whisper
moviepy
imageio_ffmpeg
pydub
uuid
langchain
openai